confirmation alert

+++ src/app/components/products/products.component.ts
@@ -4,6 +4,8 @@ import { DataService } from '../../services/data.service';
 import { Observable } from 'rxjs';
 import { Product } from '../../shared/product';
 import { Router } from '@angular/router';
+import { MatDialog, MatSnackBar } from '@angular/material';
+import { AlertComponent } from '../alert/alert.component';
 
 @Component({
   selector: 'app-products',
@@ -21,7 +23,11 @@ export class ProductsComponent implements OnInit {
   dataSource: Observable<Product[]>;
   columnsToDisplay = ['id', 'Name', 'SKU', 'Date', 'Price'];
   expandedElement: Product | null;
-  constructor(private dataservice: DataService, private router: Router) {  }
+  constructor(
+    private dataservice: DataService,
+    private router: Router,
+    private snackBar: MatSnackBar,
+    private dialog: MatDialog) {  }
 
   ngOnInit() {
     this.dataSource = this.dataservice.getProducts();
@@ -32,13 +38,28 @@ export class ProductsComponent implements OnInit {
   }
 
   onDelete(id: number) {
-    this.dataservice.deleteProduct(id).subscribe((data) => {
-      console.log(data);
-      this.dataSource = this.dataservice.getProducts();
-      this.router.navigate(['/']);
-    })
+    const dialogRef = this.dialog.open(AlertComponent,{
+      data:{
+        message: 'Are you sure want to delete?',
+        buttonText: {
+          ok: 'Save',
+          cancel: 'No'
+        }
+      }
+    });
+
+    dialogRef.afterClosed().subscribe((confirmed: boolean) => {
+      if (confirmed) {
+        this.dataservice.deleteProduct(id).subscribe((data) => {
+          console.log(data);
+          this.dataSource = this.dataservice.getProducts();
+          this.router.navigate(['/']);
+        });
+        this.snackBar.open('Okay the product is deleted', 'close', {
+          duration: 3000,
+        });
+      }
+    });
   }
-
-
 }
